侧边栏 宽度 根据屏幕宽度调整

+++ app/src/main/java/com/loyo/oa/v2/activity/home/bean/HttpAchieves.java
@@ -9,5 +9,5 @@ import java.io.Serializable;
 public class HttpAchieves implements Serializable {
     public int achieveType;
     public int achieveMoney;
-    public int finshMoney;
+    public long finshMoney;
 }
